Add support for returnMap option
Makes transpilation as part of a bundling process more efficient, by making
the source map data available as an object to be consumed by the next stage
in the chain. Previously, the source map was always encoded to a JS comment
containing a data: URI, which would need to be re-decoded and parsed by
the next step in the chain each time.